+++ CHANGES.md
@@ -6,6 +6,7 @@
 * Add support for names to Floating IP commands
 * Make `--ip-range` on `hcloud network add-subnet` optional
 * Add debug mode (use `HCLOUD_DEBUG` and `HCLOUD_DEBUG_FILE`)
+* Add `hcloud server ip` command
 
 ## v1.13.0